+++ testkit/tests/inflating_currency/proto/mod.rs
@@ -1,0 +1,28 @@
+// Copyright 2018 The Exonum Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Module of the rust-protobuf generated files.
+
+// For protobuf generated files.
+#![allow(bare_trait_objects)]
+#![allow(renamed_and_removed_lints)]
+
+pub use self::inflating_cryptocurrency::{TxCreateWallet, TxTransfer, Wallet};
+
+include!(concat!(
+    env!("OUT_DIR"),
+    "/currency_example_protobuf_mod.rs"
+));
+
+use exonum::encoding::protobuf::*;