+++ testkit/tests/service_hooks/hooks.rs
@@ -14,6 +14,7 @@
 
 //! A special service which generates transactions on `after_commit` events.
 
+use super::proto;
 use exonum::{
     blockchain::{
         ExecutionResult, Service, ServiceContext, Transaction, TransactionContext, TransactionSet,
@@ -27,13 +28,21 @@ use exonum::{
 
 pub const SERVICE_ID: u16 = 512;
 
-transactions! {
-    pub HandleCommitTransactions {
+#[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, PartialEq)]
+#[exonum(pb = "proto::TxAfterCommit")]
+pub struct TxAfterCommit {
+    pub height: Height,
+}
 
-        struct TxAfterCommit {
-            height: Height,
-        }
+impl TxAfterCommit {
+    pub fn new(height: Height) -> Self {
+        Self { height }
     }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, TransactionSet)]
+pub enum HandleCommitTransactions {
+    TxAfterCommit(TxAfterCommit),
 }
 
 impl Transaction for TxAfterCommit {